+++ blocks/identity-block/components/SocialSignOn/default.test.jsx
@@ -1,21 +1,11 @@
 import React from 'react';
 import { mount, shallow } from 'enzyme';
-import { act } from 'react-dom/test-utils';
 import SocialSignOn from './index';
 import useIdentity from '../Identity';
 
 jest.mock('../Identity');
 
 describe('Subscriptions Social Login Component', () => {
-  it('renders nothing if identity not initialized', () => {
-    useIdentity.mockImplementation(() => ({ isInitialized: false }));
-
-    const uninitializedWrapper = shallow(
-      <SocialSignOn onError={() => null} redirectURL="#" />,
-    );
-    expect(uninitializedWrapper.html()).toBe(null);
-  });
-
   it('renders nothing if config settings are false', () => {
     useIdentity.mockImplementation(() => ({
       isInitialized: true,
@@ -25,13 +15,17 @@ describe('Subscriptions Social Login Component', () => {
           googleClientId: false,
           facebookAppId: false,
         },
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
       },
     }));
 
     const wrapper = shallow(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
-    expect(wrapper.html()).toBe(null);
+    // don't render any facebook stuff, only show wrapper
+    expect(wrapper.html()).toBe('<section class="xpmedia-social-signin-wrapper"></section>');
   });
 
   it('renders only Google button', () => {
@@ -43,10 +37,13 @@ describe('Subscriptions Social Login Component', () => {
           googleClientId: true,
           facebookAppId: false,
         },
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
       },
     }));
 
-    const wrapper = shallow(
+    const wrapper = mount(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
     expect(wrapper.find('#google-sign-in-button')).toHaveLength(1);
@@ -62,10 +59,13 @@ describe('Subscriptions Social Login Component', () => {
           googleClientId: false,
           facebookAppId: true,
         },
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
       },
     }));
 
-    const wrapper = shallow(
+    const wrapper = mount(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
     expect(wrapper.find('#google-sign-in-button')).toHaveLength(0);
@@ -82,12 +82,9 @@ describe('Subscriptions Social Login Component', () => {
           signinRecaptcha: false,
           recaptchaSiteKey: '6LdXKVQcAAAAAO2tv3GdUbSK-1vcgujX6cP0IgF_',
         })),
-        initGoogleLogin: jest.fn(async () => Promise.resolve({
-          init: true,
-        })),
-        initializeFacebook: jest.fn(async () => Promise.resolve({
-          init: true,
-        })),
+        initGoogleLogin: () => {},
+        initFacebookLogin: () => {},
+        initializeFacebook: () => {},
         configOptions: {
           googleClientId: true,
           facebookAppId: true,
@@ -95,14 +92,13 @@ describe('Subscriptions Social Login Component', () => {
       },
     }));
 
-    const wrapper = shallow(
+    const wrapper = mount(
       <SocialSignOn onError={() => null} redirectURL="#" />,
     );
     expect(wrapper.html()).not.toBe(null);
   });
 
   it('renders placeholders for Google and Facebook sign-in buttons', () => {
-    let wrapper;
     useIdentity.mockImplementation(() => ({
       isInitialized: true,
       isLoggedIn: () => true,
@@ -124,12 +120,10 @@ describe('Subscriptions Social Login Component', () => {
         },
       },
     }));
-    act(() => {
-      wrapper = mount(
-        <SocialSignOn onError={() => null} redirectURL="#" />,
-      );
-      expect(wrapper.find('#google-sign-in-button')).toHaveLength(1);
-      expect(wrapper.find('.fb-login-button')).toHaveLength(1);
-    });
+    const wrapper = mount(
+      <SocialSignOn onError={() => null} redirectURL="#" />,
+    );
+    expect(wrapper.find('#google-sign-in-button')).toHaveLength(1);
+    expect(wrapper.find('.fb-login-button')).toHaveLength(1);
   });
 });